requirements.txt
python-telegram-bot==20.7
aiohttp==3.9.5
python-dotenv==1.0.1
openai==1.30.1
APScheduler==3.10.4
tzlocal==5.2
